Initial refactoring
Moved Firebase DB work into the main file, and made up some pseudo-routing.
